Refactor into a module, all tests pass
