+++ src/f4enix/output/rssa/__init__.py
@@ -1,0 +1,4 @@
+from f4enix.output.rssa.rssa import RSSA
+from f4enix.output.rssa.rssa_plotting import PlotParameters, SpectraPlotParameters
+
+__all__ = ["RSSA", "PlotParameters", "SpectraPlotParameters"]
